add picker view and locale provider web suffix
